requirements.txt
Django==5.0.6
djangorestframework==3.15.2
gunicorn==22.0.0
psycopg2==2.9.9
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.27
django-ckeditor==6.7.0
